+++ src/mdf_viewer/view/main_window.py
@@ -2197,6 +2197,7 @@ class MainWindow(QMainWindow):
                 self._controller.refresh_z_order()
                 self._controller.refresh_display_names()
                 self._controller.refresh_plot_background()
+                self._controller.refresh_signal_browser_view_mode()
                 for table in self._all_active_signals_tables():
                     table.set_shorten_names_enabled(self._settings.display_name_rule_enabled)